+++ src/SecretObject.php
@@ -12,6 +12,7 @@ use MagicObject\Util\PicoGenericObject;
 use MagicObject\Util\PicoIniUtil;
 use MagicObject\Util\PicoStringUtil;
 use MagicObject\Util\PicoYamlUtil;
+use MagicObject\Util\ValidationUtil;
 use ReflectionClass;
 use stdClass;
 
@@ -1227,6 +1228,32 @@ class SecretObject extends stdClass // NOSONAR
         $snake = $this->_snakeYaml();
         $input = $this->valueArray($snake);
         return PicoYamlUtil::dump($input, $inline, $indent, $flags);
+    }
+
+    /**
+     * Validate the current object based on property annotations.
+     *
+     * This method checks the properties of the current object against validation annotations.
+     * If any validation rule fails, an InvalidValueException will be thrown.
+     *
+     * @param string|null $parentPropertyName The name of the parent property, if applicable (for nested validation).
+     * @param array|null $messageTemplate Optional custom message templates for validation errors.
+     * @param MagicObject|null $reference Optional reference object. If provided and is an instance of MagicObject,
+     *                                    validation will use the property annotations from the reference class
+     *                                    (not from the validated object's class), but the data to validate is taken from the current object.
+     * @throws InvalidValueException If validation fails.
+     * @return self Returns the current instance for method chaining.
+     */
+    public function validate($parentPropertyName = null, $messageTemplate = null, $reference = null)
+    {
+        $objectToValidate = $this;
+        if (isset($reference) && $reference instanceof MagicObject) {
+            // If a reference object is provided, validation will use the property annotations
+            // from the reference class, but the data to validate is from the current object.
+            $objectToValidate = $reference->loadData($this);
+        }
+        ValidationUtil::getInstance($messageTemplate)->validate($objectToValidate, $parentPropertyName);
+        return $this;
     }
 
     /**